pyproject.toml
[project]
name = "sales-assist-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "db-dtypes>=1.5.0",
    "google-adk>=1.24.1",
    "google-cloud-aiplatform[agent-engines]>=1.136.0",
    "google-cloud-bigquery>=3.40.0",
    "pandas>=2.3.3",
    "python-dotenv>=1.2.1",
]
